polars_dtype

Module categorical

Structs

CategoricalMapping

Categories

A (named) object which is used to indicate which categorical data types have the same mapping.

FrozenCategories

An ordered collection of unique strings with an associated pre-computed mapping to go from string <-> index.

Enums

CategoricalPhysical

The physical datatype backing a categorical / enum.

Traits

CatNative

Functions

ensure_same_categories

ensure_same_frozen_categories

Type Aliases

CatSize